+++ app/test/foliate_session_test.dart
@@ -34,6 +34,8 @@ void main() {
     expect(command['type'], 'open');
     expect(command['cfi'], session.currentCfi);
     expect(command['pageCount'], session.pageCount);
+    expect(command['fontSize'], 18);
+    expect(FoliateBridge.openSession(session, fontSize: 22)['fontSize'], 22);
     expect(command.containsKey('FoliateView'), isFalse);
   });